Do not export PortalProvider
