+++ packages/mui/src/Root.internal.tsx
@@ -4,8 +4,15 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as React from "react";
+import { PortalContext as AkPortalContext } from "@ariakit/react/portal";
+import { Role } from "@ariakit/react/role";
 import createCache from "@emotion/cache";
 import { CacheProvider } from "@emotion/react";
+import {
+	PortalContext,
+	RootContext,
+} from "@stratakit/foundations/secret-internals";
+import { useSafeContext } from "@stratakit/internal-utils/hooks";
 
 import type { StylisElement } from "@emotion/cache";
 
@@ -85,3 +92,29 @@ function hash(value: string, length: number) {
 				(value.charCodeAt(3) | 0)
 		: 0;
 }
+
+// ----------------------------------------------------------------------------
+
+type PortalContextValue = React.ContextType<typeof PortalContext>;
+
+/**
+ * Provides the element that will contain the portal.
+ */
+export function PortalProvider(
+	props: React.PropsWithChildren<PortalContextValue>,
+) {
+	const { children, container, unstable_getContainer } = props;
+
+	const { portalProvider } = useSafeContext(RootContext);
+
+	return (
+		<PortalContext.Provider value={{ container, unstable_getContainer }}>
+			<AkPortalContext.Provider value={container}>
+				<Role render={portalProvider ?? <React.Fragment />}>{children}</Role>
+			</AkPortalContext.Provider>
+		</PortalContext.Provider>
+	);
+}
+DEV: PortalProvider.displayName = "PortalProvider";
+
+// ----------------------------------------------------------------------------
